+++ AddOns/BetterBags/CHANGELOG.md
@@ -1,6 +1,11 @@
 # BetterBags
 
-## [v0.0.90](https://github.com/Cidan/BetterBags/tree/v0.0.90) (2024-02-05)
-[Full Changelog](https://github.com/Cidan/BetterBags/compare/v0.0.89...v0.0.90) [Previous Releases](https://github.com/Cidan/BetterBags/releases)
+## [v0.0.91](https://github.com/Cidan/BetterBags/tree/v0.0.91) (2024-02-06)
+[Full Changelog](https://github.com/Cidan/BetterBags/compare/v0.0.90...v0.0.91) [Previous Releases](https://github.com/Cidan/BetterBags/releases)
 
-- Added a small work around in the case of missing currencies  
+- Classic Support (#92)  
+    * Added support for Classic  
+    * Added small animations to currency and bag slot frames  
+    * Change some sound effects  
+    * Fixed a few grid alignment issues  
+- Add subtype filter to the categories (#89)